Fix frameLimiterPattern offset and add audioLimiter2Pattern.
frameLimiterPattern was writing to an offset of +40 to the pattern, instead of +20. Ironically this prevented a crash when reloading a savefile, because it would fail to match the pattern for the already nop'ed code.

Added a new audio engine limiter, found by Special For. I'm honestly not sure how this one works, seems like it should also limit to 188 fps but it does so at 3x more - 564. Regardless, noping a function call removes that sleep too.